requirements.txt
opencv-python
gunicorn
tensorflow
keras
flask
Pillow
numpy